Move Discovery.Config into Akka.Management project
