--- src/discovery/config/Akka.Discovery.Config.Hosting/Akka.Discovery.Config.Hosting.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-    </PropertyGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\..\management\Akka.Management\Akka.Management.csproj" />
-    </ItemGroup>
-
-</Project>
