+++ src/management/Akka.Management.Tests/Akka.Management.Tests.csproj
@@ -6,6 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Akka.Hosting.TestKit" Version="$(AkkaHostingVersion)" />
     <PackageReference Include="Akka.Cluster.Hosting" Version="$(AkkaHostingVersion)" />
     <PackageReference Include="Akka.Remote.Hosting" Version="$(AkkaHostingVersion)" />
     <PackageReference Include="FluentAssertions" Version="$(FluentAssertionVersion)" />
