+++ src/management/Akka.Management/Discovery/Config/Service.cs
@@ -10,6 +10,7 @@ using System.Text;
 using Akka.Configuration;
 using Akka.Hosting;
 
+// ReSharper disable once CheckNamespace
 namespace Akka.Discovery.Config.Hosting;
 
 public class Service
